backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-bigquery==3.13.0
python-dotenv==1.0.0
redis==5.0.1
h3==3.7.6
pydantic==2.5.2
pydantic-settings==2.1.0
pandas==2.1.3
db-dtypes==1.2.0
